feat: Enhance product management and authentication features
- Added mongoose-aggregate-paginate-v2 for pagination support in product queries.
- Implemented product addition and retrieval by seller in product.controller.js.
- Created product.routes.js for handling product-related routes.
- Updated auth.controller.js to improve user authentication responses.
- Introduced error middleware for structured error handling.
- Enhanced JWT verification in auth.middleware.js for better error management.
- Updated category and product models with new fields and indexing.
- Created address.routes.js and auth.routes.js for address and authentication routes.

+++ server/src/utils/cloudinary.js
@@ -1,4 +1,4 @@
-import {v2 as cloudinary} from "cloudinary";
+import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
 
 cloudinary.config({
@@ -7,26 +7,30 @@ cloudinary.config({
     api_secret: process.env.CLOUDINARY_API_SECRET
 })
 
-// function for uploading files such as videos, images or pdf on cloud with cloudinary 
-// service
 const uploadOnCloudinary = async (localFilePath) => {
     try {
-        if (!localFilePath) return null
-        //upload the file on cloudinary
+        if (!localFilePath) return null;
+        console.log("Path:", localFilePath);
         const response = await cloudinary.uploader.upload(localFilePath, {
-            resource_type: "auto"
-        })
-        // file has been uploaded successfull
-        //console.log("file is uploaded on cloudinary ", response.url);
-        fs.unlinkSync(localFilePath) // delete the local file after upload 
+            resource_type: "image",
+        });
+
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
+
         return response;
 
     } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
+        console.error("Cloudinary Error:", error);
+
+        if (localFilePath && fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
+
         return null;
     }
-}
-
-export {uploadOnCloudinary}
+};
 
 
+export { uploadOnCloudinary };
